+++ js/aurelia/services/foundation/adapter.js
@@ -1,0 +1,18 @@
+class FoundationAdapter {
+  constructor(FoundationApi: foundationApi) {
+    var service    = {};
+
+    service.activate = activate;
+    service.deactivate = deactivate;
+
+    return service;
+
+    function activate(target) {
+      foundationApi.publish(target, 'show');
+    }
+
+    function deactivate(target) {
+      foundationApi.publish(target, 'hide');
+    }
+  }
+}