+++ js/aurelia/services/foundation/setup.js
@@ -1,0 +1,13 @@
+class Setup {
+  constructor() {
+    // Attach FastClick
+    if (FastClick) {
+      FastClick.attach(document.body);
+    }
+
+    // Attach viewport units buggyfill
+    if (viewportUnitsBuggyfill) {
+      viewportUnitsBuggyfill.init();
+    }
+  }
+}