+++ js/aurelia/services/foundation/utils.js
@@ -1,0 +1,20 @@
+class Utils {
+  constructor() {
+    this.throttle = throttleUtil;
+  }
+
+  throttleUtil(func, delay) {
+    var timer = null;
+
+    return function () {
+      var context = this, args = arguments;
+
+      if (timer === null) {
+        timer = setTimeout(function () {
+          func.apply(context, args);
+          timer = null;
+        }, delay);
+      }
+    };
+  }
+}