+++ js/aurelia/services/media-query/mediaquery.js
@@ -1,0 +1,7 @@
+import Core from './core'
+
+class MediaQuery {
+  constructor(MQInit: mqInit, MQHelpers: mqHelpers, MQ: mq) {
+    mqInit.init();
+  }
+}