+++ js/aurelia/services/routing/state.js
@@ -1,0 +1,7 @@
+@inject StateProvider, Router;
+class FoundationState
+  constructor(StateProvider: stateProvider, Router: router) {
+    var complexViews = {};
+    this.registerDynamicRoutes = registerDynamicRoutes(router.routes(), complexViews);
+  }
+}